+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - MPEG-4 Part 14 Audio (MP4)
 - MPEG-4 Part 14 Subtitles (MP4)
 - Matroska Subtitles (MKS)
+- Silicon Graphics Movie (SGI)
 
 # Version 0.20.0 (2023-09-07)